Fixed a warning.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/apple-local-200502-branch@102312 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/print-tree.c
@@ -344,7 +344,7 @@ print_node (FILE *file, const char *prefix, tree node, int indent)
         fputs (" asm-function", file);
       if (DECL_CW_ASM_NORETURN (node))
         fputs (" asm-noreturn", file);
-      if (DECL_CW_ASM_FRAME_SIZE (node) != -2)
+      if (DECL_CW_ASM_FRAME_SIZE (node) != (unsigned int)-2)
         fprintf (file, " asm-frame-size %d", DECL_CW_ASM_FRAME_SIZE (node));
       /* APPLE LOCAL end CW asm blocks */
       if (TREE_CODE (node) == TYPE_DECL && TYPE_DECL_SUPPRESS_DEBUG (node))
